fix: move trial expiry date from summary to settings api
