+++ src/common/meta/organization.rs
@@ -147,7 +147,6 @@ pub struct OrgSummary {
     pub alerts: AlertSummary,
     pub total_functions: i64,
     pub total_dashboards: i64,
-    pub trial_period_expiry: Option<i64>,
 }
 
 #[derive(Default, Serialize, Deserialize, ToSchema)]
@@ -266,6 +265,10 @@ pub struct OrganizationSetting {
     pub enable_streaming_search: bool,
     #[serde(default = "default_auto_refresh_interval")]
     pub min_auto_refresh_interval: u32,
+    // we skip this as this is actually stored in another table
+    // and only applicable for cloud
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub free_trial_expiry: Option<i64>,
 }
 
 impl Default for OrganizationSetting {
@@ -278,6 +281,7 @@ impl Default for OrganizationSetting {
             enable_websocket_search: default_enable_websocket_search(),
             enable_streaming_search: default_enable_streaming_search(),
             min_auto_refresh_interval: default_auto_refresh_interval(),
+            free_trial_expiry: None,
         }
     }
 }
